+++ .babelrc
@@ -1,3 +1,3 @@
 {
-    "presets": ["modern-browsers"]
+    "plugins": [["@babel/plugin-proposal-optional-chaining"]]
 }
