+++ package.json
@@ -9,8 +9,35 @@
   "jsdelivr": "dist/immer.umd.js",
   "jsnext:main": "dist/immer.module.js",
   "react-native": "dist/immer.module.js",
-  "source": "src/index.ts",
-  "types": "./dist/index.d.ts",
+  "source": "src/immer.ts",
+  "types": "./dist/immer.d.ts",
+  "sideEffects": false,
+  "exports": {
+    ".": {
+      "browser": "./dist/immer.module.js",
+      "umd": "./dist/immer.umd.js",
+      "import": "./dist/immer.module.js",
+      "require": "./didst/immer.js"
+    },
+    "es5": {
+      "browser": "./dist/es5.module.js",
+      "umd": "./dist/es5.umd.js",
+      "import": "./dist/es5.module.js",
+      "require": "./didst/es5.js"
+    },
+    "mapset": {
+      "browser": "./dist/mapset.module.js",
+      "umd": "./dist/mapset.umd.js",
+      "import": "./dist/mapset.module.js",
+      "require": "./didst/mapset.js"
+    },
+    "patches": {
+      "browser": "./dist/patches.module.js",
+      "umd": "./dist/patches.umd.js",
+      "import": "./dist/patches.module.js",
+      "require": "./didst/patches.js"
+    }
+  },
   "scripts": {
     "test": "jest && yarn-or-npm test:build && yarn-or-npm test:flow",
     "test:perf": "NODE_ENV=production yarn-or-npm build && cd __performance_tests__ && babel-node add-data.js && babel-node todo.js && babel-node incremental.js",
@@ -19,7 +46,9 @@
     "watch": "jest --watch",
     "coverage": "jest --coverage",
     "coveralls": "jest --coverage && cat ./coverage/lcov.info | ./node_modules/.bin/coveralls && rm -rf ./coverage",
-    "build": "rimraf dist/ && yarn-or-npm bili && yarn-or-npm typed",
+    "build": "rimraf dist/ && yarn build:immer && yarn build:es5 && yarn build:patches && yarn build:mapset",
+    "build:immer": "microbundle build --raw",
+    "build:es5": "microbundle build --raw --entry src/plugins/es5.ts --output dist",
     "typed": "cpx 'src/immer.js.flow' dist -v",
     "publish-docs": "cd website && GIT_USER=mweststrate USE_SSH=true yarn run publish-gh-pages",
     "start": "cd website && yarn start"
@@ -53,38 +82,27 @@
     "src"
   ],
   "devDependencies": {
-    "@babel/cli": "^7.2.3",
-    "@babel/core": "^7.3.4",
-    "@babel/node": "^7.2.2",
-    "@babel/plugin-external-helpers": "^7.2.0",
-    "@babel/preset-env": "^7.3.4",
-    "@types/jest": "^24.0.11",
-    "babel-jest": "^24.4.0",
-    "babel-preset-modern-browsers": "^13.1.0",
-    "bili": "^4.8.1",
+    "@babel/plugin-proposal-optional-chaining": "^7.8.3",
     "coveralls": "^3.0.0",
     "cpx": "^1.5.0",
     "cross-env": "^5.1.3",
     "deep-freeze": "^0.0.1",
-    "expect": "^24.7.1",
     "flow-bin": "^0.68.0",
     "husky": "^1.2.0",
     "immutable": "^3.8.2",
-    "jest": "^24.7.1",
+    "jest": "^25.1.0",
     "lodash": "^4.17.4",
     "lodash.clonedeep": "^4.5.0",
+    "microbundle": "^0.12.0-next.8",
     "prettier": "1.19.1",
     "pretty-quick": "^1.8.0",
     "redux": "^4.0.5",
-    "regenerator-runtime": "^0.11.1",
     "rimraf": "^2.6.2",
-    "rollup-plugin-typescript2": "^0.25.3",
     "seamless-immutable": "^7.1.3",
     "semantic-release": "^17.0.2",
     "spec.ts": "^1.1.0",
-    "ts-jest": "^24.2.0",
-    "typescript": "^3.7.3",
-    "yarn-or-npm": "^2.0.4"
+    "ts-jest": "^25.2.0",
+    "typescript": "^3.7.3"
   },
   "jest": {
     "moduleFileExtensions": [
