+++ src/internal.ts
@@ -5,5 +5,5 @@ export * from "./common"
 export * from "./scope"
 export * from "./finalize"
 export * from "./proxy"
-export * from "./immer"
+export * from "./immerClass"
 export * from "./plugins"
